Create schema ui handler and separate from form builder.
