+++ modules/json_form_widget/json_form_widget.services.yml
@@ -1,7 +1,7 @@
 services:
   json_form.builder:
     class: \Drupal\json_form_widget\JsonFormBuilder
-    arguments: ['@dkan.metastore.schema_retriever','@json_form.string_helper','@json_form.object_helper', '@json_form.array_helper']
+    arguments: ['@dkan.metastore.schema_retriever','@json_form.string_helper','@json_form.object_helper', '@json_form.array_helper', '@json_form.schema_ui_handler']
   json_form.value_handler:
     class: \Drupal\json_form_widget\JsonFormValueHandler
   json_form.string_helper:
@@ -11,3 +11,6 @@ services:
   json_form.array_helper:
     class: \Drupal\json_form_widget\JsonFormArrayHelper
     arguments: ['@json_form.object_helper']
+  json_form.schema_ui_handler:
+    class: \Drupal\json_form_widget\JsonFormSchemaUiHandler
+    arguments: ['@dkan.metastore.schema_retriever', '@logger.factory']
